ByeSPACE - Final project
A Haskell interpreter for a simple programming language. It uses a Parser defined as an instance of Functor, Monad, Applicative and Alternative, and an Environment which is the "memory" of the program. It can handle arithmetic expressions, boolean expressions, while loops, if then else command, the ternary operator, and also some array operations as concatenation.
(Note: in array handling, it is possible to define an element of the array as an arithmetic expression)